Refactor attestation modules and add comprehensive documentation
- Move none attestation implementation to dedicated module
- Move packed attestation implementation to dedicated module
- Add detailed documentation for none attestation format
- Add detailed documentation for packed attestation format
- Improve documentation for TPM attestation implementation
- Add comprehensive error handling documentation
- Update core attestation module to use new module structure
- Add constants and helper functions to utils module

The refactoring improves code organization and maintainability while
adding comprehensive documentation that covers all attestation formats
supported by the library.
