+++ oauth2_passkey/src/passkey/main/attestation/mod.rs
@@ -1,4 +1,6 @@
 mod core;
+mod none;
+mod packed;
 mod tpm;
 mod utils;
